refactor: Move `config/` into `data/`
