+++ src/main/java/io/github/null2264/cobblegen/CobbleGen.java
@@ -4,7 +4,7 @@ import com.mojang.brigadier.CommandDispatcher;
 import com.mojang.brigadier.builder.LiteralArgumentBuilder;
 import io.github.null2264.cobblegen.compat.LoaderCompat;
 import io.github.null2264.cobblegen.compat.TextCompat;
-import io.github.null2264.cobblegen.config.ConfigMetaData;
+import io.github.null2264.cobblegen.data.config.ConfigMetaData;
 import io.github.null2264.cobblegen.data.model.CGRegistry;
 import io.github.null2264.cobblegen.util.CGLog;
 import net.minecraft.commands.CommandSourceStack;
@@ -14,7 +14,7 @@ import org.jetbrains.annotations.ApiStatus;
 import java.io.File;
 import java.nio.file.Path;
 
-import static io.github.null2264.cobblegen.config.ConfigHelper.loadConfig;
+import static io.github.null2264.cobblegen.data.config.ConfigHelper.loadConfig;
 
 //#if FORGE>=1
     //#if FORGE==2
@@ -28,8 +28,6 @@ public class CobbleGen implements net.fabricmc.api.ModInitializer
 //#endif
 {
     public static final String MOD_ID = "cobblegen";
-    public static final ResourceLocation SYNC_CHANNEL = new ResourceLocation(MOD_ID, "sync");
-    public static final ResourceLocation SYNC_PING_CHANNEL = new ResourceLocation(MOD_ID, "ping");
     /**
      * @deprecated Now only for internal usage. Use the parameter {@link CGRegistry registry} instead
      */
