--- src/main/java/io/github/null2264/cobblegen/config/Config.java
@@ -1,3 +1,0 @@
-package io.github.null2264.cobblegen.config;
-
-public interface Config {}
