+++ src/main/java/io/github/null2264/cobblegen/data/config/Config.java
@@ -1,0 +1,3 @@
+package io.github.null2264.cobblegen.data.config;
+
+public interface Config {}
