+++ src/main/java/io/github/null2264/cobblegen/data/config/CustomGen.java
@@ -1,4 +1,4 @@
-package io.github.null2264.cobblegen.config;
+package io.github.null2264.cobblegen.data.config;
 
 import blue.endless.jankson.JsonObject;
 import blue.endless.jankson.annotation.Deserializer;
@@ -10,7 +10,7 @@ import org.jetbrains.annotations.Nullable;
 import java.util.List;
 import java.util.Map;
 
-import static io.github.null2264.cobblegen.config.ConfigHelper.generatorFromJson;
+import static io.github.null2264.cobblegen.data.config.ConfigHelper.generatorFromJson;
 import static io.github.null2264.cobblegen.util.Constants.JANKSON;
 
 public class CustomGen implements JanksonSerializable
