+++ src/main/java/io/github/null2264/cobblegen/data/model/BuiltInGenerator.java
@@ -1,6 +1,6 @@
 package io.github.null2264.cobblegen.data.model;
 
-import io.github.null2264.cobblegen.config.WeightedBlock;
+import io.github.null2264.cobblegen.data.config.WeightedBlock;
 import io.github.null2264.cobblegen.data.CGIdentifier;
 import io.github.null2264.cobblegen.util.Util;
 import lombok.val;
